refactor: move board domain to features/board/
