+++ src/features/board/index.ts
@@ -1,0 +1,4 @@
+export * from "./components/KanbanBoard";
+export * from "./context/KanbanContext";
+export * from "./hooks/useKanban";
+export * from "./types/board.types";
