+++ src/features/board/types/board.types.ts
